+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+* Add pycryptodomex version requiremenet range (`>=3.9,<4`)
+
 ## v0.4.5 - October 24, 2020
 * Fix loading RomFS from a filename in `RomFSReader`
 * Fix loading ExeFS from a filename in `RomFSReader`